feat(react-ui): rename dialog to modal
