+++ libs/react-ui/src/index.ts
@@ -4,4 +4,4 @@ export {
   ButtonSize,
 } from './lib/button/button';
 
-export { default as Dialog } from './lib/Dialog';
+export { default as Modal } from './lib/Modal';
